chore: improve CheckboxGroup exports
